handle for todo case
