+++ src/main/java/Tasks/Deadline.java
@@ -1,4 +1,6 @@
-public class Deadline extends Task  {
+package Tasks;
+
+public class Deadline extends Task {
     protected String by;
 
     public Deadline(String description, String by) {
