+++ src/main/java/Tasks/Event.java
@@ -1,4 +1,6 @@
-public class Event extends Task{
+package Tasks;
+
+public class Event extends Task {
     protected String at;
 
     public Event(String description, String at) {
