+++ src/main/java/Tasks/Task.java
@@ -1,3 +1,7 @@
+package Tasks;
+
+import java.util.ArrayList;
+
 public class Task {
     protected String description;
     protected boolean isDone;
@@ -13,6 +17,15 @@ public class Task {
 
     public void setDone(){
         this.isDone = true;
+    }
+    public static void markAsDone(int arrayIndex) {
+        ArrayList<Task> tasks = new ArrayList<>();
+        Task task = tasks.get(arrayIndex);
+        task.setDone();
+        System.out.println("     Nice! I've marked this task as done:\n"
+                + task
+                + "\n"
+                + "___________________________________________________________________\n");
     }
 
     @Override
