+++ src/main/java/Tasks/ToDo.java
@@ -1,4 +1,6 @@
-public class ToDo extends Task{
+package Tasks;
+
+public class ToDo extends Task {
 
     public ToDo(String description){
         super(description);
